General Recipe I/O
Recipe for checking resource state of matter in addition to name.
Implemented and tested in Conveyor.
