+++ Assets/Audio/Downloaded/ES_Doo Dat Wah - Jules Gaia.zip.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bfbdca6f2bdc19e4db0c70e35aa7b7bd
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
