+++ Assets/Audio/Downloaded/ES_Just Playing - Jules Gaia.zip.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7766b87e774bcb443b3a06ce259cd235
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
