+++ Assets/Audio/Downloaded/ES_Midnight Swing - Jules Gaia.zip.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9d7561bf81871ff4d85d216c8b366514
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
