+++ Assets/Audio/Downloaded/ES_Shake Down - Jules Gaia.zip.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 28db2cb03bedab543851937d6aeb29fe
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
