+++ Assets/Prefabs/Machines.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3c447e14170d9ae44931d51d3369934e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
