+++ Assets/Prefabs/Machines/Conveyor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: eb894222aba72884aa9d97040fe065a2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
