+++ Assets/ResourceMap/Recipes/ConveyorRecipe.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2246790969471609759}
-  - component: {fileID: 8140615625731595231}
+  - component: {fileID: -7420448965816891819}
   m_Layer: 0
   m_Name: ConveyorRecipe
   m_TagString: Untagged
@@ -31,7 +31,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &8140615625731595231
+--- !u!114 &-7420448965816891819
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -40,7 +40,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6185347371653292868}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b4bba6a219c413d438a29805e734fc66, type: 3}
+  m_Script: {fileID: 11500000, guid: 04769c8603654cf49c1d75a05e60b1b0, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   InResources:
@@ -49,4 +49,4 @@ MonoBehaviour:
   OutResources:
   - resource: {fileID: 4724451604307785483, guid: f823c0e5707e96442895c2dd74d7db58, type: 3}
     num: 1
-  ticks: 1
+  ticks: 0
