--- Assets/Scripts/CameraFollow.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 245c4a9065d74ffd9f1e257ee04d12ac
-timeCreated: 1633650789
