+++ Assets/Scripts/Conductor.cs
@@ -76,7 +76,6 @@ public class Conductor : MonoBehaviour {
     public void Tick() {
         foreach (Machine machine in _allMachines) {
             if (machine.OutputMachines.Count == 0) {
-                print("Root tick is " + machine.name);
                 machine.Tick();
             }
         }
