--- Assets/Scripts/Conveyor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6e08f96cbe424c34a40c47fe9e1dbc09
-timeCreated: 1633657748
