+++ Assets/Scripts/Graphics/SmoothSprite.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ca8f9730e92c0943ab11ca8b3fde38d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
