--- Assets/Scripts/Recipe.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4bba6a219c413d438a29805e734fc66
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
