--- Assets/Scripts/Resource.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-[Serializable]
-public class Resource : MonoBehaviour {
-    public int price;
-    public string ResourceName;
-    public SmoothSprite MySmoothSprite { get; private set; }
-
-    void Start() {
-        MySmoothSprite = GetComponent<SmoothSprite>();
-    }
-}
