+++ Assets/Scripts/ResourceHierarchy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1593e1f25f277334087b0efb4a458827
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
