+++ Assets/Scripts/ResourceHierarchy/ExactRecipe.cs
@@ -1,0 +1,5 @@
+﻿public class ExactRecipe : Recipe {
+    protected override bool _isInputValidR(Resource recipeResource, Resource compareAgainst) {
+        return recipeResource.ResourceName == compareAgainst.ResourceName;
+    }
+}
