+++ Assets/Scripts/ResourceHierarchy/MatterRecipe.cs
@@ -1,0 +1,5 @@
+﻿public class MatterRecipe : Recipe {
+    protected override bool _isInputValidR(Resource recipeResource, Resource compareAgainst) {
+        return recipeResource.matterState == compareAgainst.matterState;
+    }
+}
