+++ Assets/Scripts/ResourceHierarchy/Resource.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5033c218530347ce92676fcadcd25fc5
+timeCreated: 1633194746
