--- Assets/Scripts/ResourceNum.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 05826c63735c4f90aa74bd524c05c0b5
-timeCreated: 1633192927
